# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kolm API",
    "version": "1.0.0",
    "description": "HTTP contract for the kolm compiler. Self-hosted (your own kolm server) and kolm.ai (managed) speak the exact same /v1 surface. JSON in, JSON out, bearer or X-Kolm-Key auth. Every successful run produces an HMAC-signed receipt over (cid, input, output, ts, k_score).",
    "termsOfService": "https://kolm.ai/terms",
    "contact": {
      "name": "kolm",
      "url": "https://kolm.ai",
      "email": "hello@kolm.ai"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "https://api.kolm.ai",
      "description": "Managed production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local dev (kolm serve)"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/v1/account": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Current account: plan, compile count, remaining quota, API key.",
        "responses": {
          "200": {
            "description": "Account snapshot.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/v1/compile": {
      "post": {
        "tags": [
          "compile"
        ],
        "summary": "Submit a compile job. Returns 202 with a compile_id and ETA.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileJobAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "402": {
            "description": "Quota exhausted on this plan.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/run": {
      "post": {
        "tags": [
          "run"
        ],
        "summary": "Run a compiled artifact against an input. Emits a signed receipt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token issued by /v1/auth/login or /v1/auth/signup."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolm-Key",
        "description": "Long-lived API key. Read from /v1/account. Use this for CI."
      }
    },
    "schemas": {
...
      "Account": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "example": "usr_01HW..."
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "team",
              "enterprise"
            ]
          },
          "compiles_count": {
            "type": "integer",
            "description": "Compiles in current period."
          },
          "compiles_remaining": {
            "type": "integer",
            "description": "Compiles left before quota or hard cap."
          },
          "api_key": {
            "type": "string",
            "example": "kolm_live_01HW...",
            "description": "Use with X-Kolm-Key header."
          }
        },
        "required": [
          "user_id",
          "email",
          "plan",
          "compiles_count",
          "compiles_remaining",
          "api_key"
        ]
      },
...
      "CompileRequest": {
        "type": "object",
        "properties": {
          "recipe": {
            "type": "object",
            "description": "Free-form recipe object: task, objective, adapter, evaluator. See /docs/manifest-v0.1.json for the canonical shape."
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Optional in-the-loop examples. Each is {input, expected}."
          },
          "base_model": {
            "type": "string",
            "example": "qwen2.5-7b-instruct",
            "description": "Base model identifier. Defaults are picked by the compile planner if omitted."
          },
          "k_min": {
            "type": "number",
            "format": "float",
            "example": 0.9,
            "description": "Hard floor: compile fails if measured K-score is below this."
          },
          "compliance_pack": {
            "type": "string",
            "nullable": true,
            "example": "hipaa-redact-v1",
            "description": "Optional pre-baked compliance pack id (HIPAA, SR 11-7, NIST AI RMF, etc.)."
          }
        },
        "required": [
          "recipe"
        ]
      },
      "CompileJobAccepted": {
        "type": "object",
        "properties": {
          "compile_id": {
            "type": "string",
            "example": "cmp_01HW..."
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running"
            ]
          },
          "eta_seconds": {
            "type": "integer",
            "description": "Estimated time-to-done."
          }
        },
        "required": [
          "compile_id",
          "status",
          "eta_seconds"
        ]
      },
...
      "RunRequest": {
        "type": "object",
        "properties": {
          "artifact_cid": {
            "type": "string",
            "nullable": true,
            "example": "cidv1:sha256:1bcf2323..."
          },
          "artifact_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "description": "Either artifact_cid OR artifact_url, not both."
          },
          "input": {
            "description": "Free-form. String for chat, object for structured tasks.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          }
        },
        "required": [
          "input"
        ]
      },
      "RunResponse": {
        "type": "object",
        "properties": {
          "output": {
            "description": "Free-form. Shape mirrors the artifact's declared output schema.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "object"
              }
            ]
          },
          "receipt": {
            "$ref": "#/components/schemas/Receipt"
          }
        },
        "required": [
          "output",
          "receipt"
        ]
      },
      "Receipt": {
        "type": "object",
        "description": "HMAC-signed receipt over (cid, input_hash, output_hash, ts).",
        "properties": {
          "cid": {
            "type": "string",
            "example": "cidv1:sha256:1bcf2323..."
          },
          "hmac": {
            "type": "string",
            "example": "hmac-sha256:9af2..."
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-15T12:34:56Z"
          },
          "k_score": {
            "type": "number",
            "format": "float",
            "example": 0.982
          },
          "input_hash": {
            "type": "string",
            "nullable": true,
            "example": "sha256:8a3d..."
          },
          "output_hash": {
            "type": "string",
            "nullable": true,
            "example": "sha256:1f2e..."
          }
        },
        "required": [
          "cid",
          "hmac",
          "ts",
          "k_score"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable error code, e.g. \"quota_exhausted\"."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail."
          },
          "trace_id": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "error",
          "message"
        ]
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid auth.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "No record for the given id or CID.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Payload failed validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}