# server.js
import 'dotenv/config';
import express from 'express';
import compression from 'compression';
import helmet from 'helmet';
import cookieParser from 'cookie-parser';
import fs from 'node:fs';
import path from 'node:path';
import { fileURLToPath } from 'node:url';
import { buildRouter } from './src/router.js';
import { provisionTenant } from './src/auth.js';
import { isProductionRuntime } from './src/env.js';
import { synthesize } from './src/synthesis.js';
import { createConcept, publishVersion } from './src/registry.js';
import { all } from './src/store.js';

const __dirname = path.dirname(fileURLToPath(import.meta.url));

const app = express();
app.disable('x-powered-by');
// Two-hop proxy chain in prod: Client → Vercel CDN → Railway edge → Express.
// `trust proxy: 2` makes express read X-Forwarded-For from the right hop so
// the rate-limit keyGenerator sees the real client IP, not Railway's edge IP.
app.set('trust proxy', isProductionRuntime() ? 2 : false);

// Security headers (S3, S4) — mounted BEFORE express.static so static
// assets get HSTS, CSP, nosniff, etc. CSP allows 'unsafe-inline' for now
// because every page still has inline <script> blocks; Sprint 1 moves
// inline scripts to /js/<page>.js and tightens CSP. 'wasm-unsafe-eval' is
// required by the on-device runtime (wllama, sqlite-vec).
app.use(helmet({
  contentSecurityPolicy: {
    directives: {
      defaultSrc: ["'self'"],
      scriptSrc: ["'self'", "'unsafe-inline'", "'wasm-unsafe-eval'", 'https://js.stripe.com', 'https://*.vercel-insights.com'],
      styleSrc: ["'self'", "'unsafe-inline'", 'https://fonts.googleapis.com'],
      imgSrc: ["'self'", 'data:', 'blob:', 'https:'],
      fontSrc: ["'self'", 'data:', 'https://fonts.gstatic.com'],
      connectSrc: ["'self'", 'https://api.anthropic.com', 'https://kolm.ai', 'https://kolmogorov-stack-production.up.railway.app', 'https://*.vercel-insights.com', 'https://api.stripe.com'],
      frameSrc: ['https://js.stripe.com'],
      workerSrc: ["'self'", 'blob:'],
      frameAncestors: ["'none'"],
      objectSrc: ["'none'"],
      baseUri: ["'self'"],
      formAction: ["'self'"],
    },
  },
  crossOriginEmbedderPolicy: false,
  crossOriginResourcePolicy: { policy: 'cross-origin' },
  strictTransportSecurity: { maxAge: 63072000, includeSubDomains: true, preload: true },
  referrerPolicy: { policy: 'strict-origin-when-cross-origin' },
  frameguard: { action: 'deny' },
  noSniff: true,
}));

// gzip everything except SSE streams (compression breaks event delivery).
app.use(compression({ filter: (req, res) => res.getHeader('Content-Type') !== 'text/event-stream' && compression.filter(req, res) }));
app.use(cookieParser());
// Stripe webhook signature verification needs the raw request body — JSON
// reparse reorders keys and breaks the HMAC. Mount express.raw() for the
// webhook route ahead of express.json() so req.body is a Buffer there only.
app.use((req, res, next) => {
  if (req.path === '/v1/stripe/webhook') {
    return express.raw({ type: '*/*', limit: '4mb' })(req, res, next);
  }
  return express.json({ limit: '4mb' })(req, res, next);
});
app.use(express.urlencoded({ extended: true }));

// /articles serves the index page directly (no 301 redirect) — must come
// BEFORE express.static so the static directory-redirect doesn't fire.
app.get('/articles', (_req, res) => {
  res.set('Cache-Control', 'public, max-age=60, must-revalidate');
  res.sendFile(path.join(__dirname, 'public', 'articles', 'index.html'));
});

// /use-cases serves the index page directly. Same reasoning — public/use-cases/
// exists as a directory, so express.static would 301-redirect /use-cases to
// /use-cases/. Pre-empt with an explicit handler.
app.get('/use-cases', (_req, res) => {
  res.set('Cache-Control', 'public, max-age=60, must-revalidate');
  res.sendFile(path.join(__dirname, 'public', 'use-cases', 'index.html'));
});

// Explicit /docs handler — serves the docs hub HTML before express.static
// can 301-redirect /docs to /docs/ (directory listing) since public/docs/
// exists as the spec-asset folder.
app.get('/docs', (_req, res) => {
  const f = path.join(__dirname, 'public', 'docs.html');
  if (fs.existsSync(f)) return res.sendFile(f);
  res.status(404).sendFile(path.join(__dirname, 'public', '404.html'));
});

// /cookbook serves cookbook.html directly. Same dance as /articles, /use-cases —
// public/cookbook/ exists as a directory of recipes, so without this hook
// express.static would 301-redirect /cookbook to /cookbook/.
app.get('/cookbook', (_req, res) => {
  res.set('Cache-Control', 'public, max-age=60, must-revalidate');
  res.sendFile(path.join(__dirname, 'public', 'cookbook.html'));
});

// RFC 9116 security.txt — serve from .well-known and as a top-level
// alias. express.static skips dot-directories on some hosts, so we serve
// explicitly to guarantee both URLs resolve.
const SECURITY_TXT = path.join(__dirname, 'public', '.well-known', 'security.txt');
for (const url of ['/.well-known/security.txt', '/security.txt']) {
  app.get(url, (_req, res) => {
    if (!fs.existsSync(SECURITY_TXT)) return res.status(404).type('text/plain').send('not found');
    res.type('text/plain; charset=utf-8');
    res.set('Cache-Control', 'public, max-age=3600');
    res.sendFile(SECURITY_TXT);
  });
}

// Static dashboard with strong caching for hashed assets, weak for HTML.
// /sdk.js gets a versioned alias (S6) — the unversioned URL stays for
// back-compat but we encourage `/sdk-<sha>.js` for SRI-pinned imports.
app.use(express.static(path.join(__dirname, 'public'), {
  setHeaders: (res, filePath) => {
    if (filePath.endsWith('.html')) res.setHeader('Cache-Control', 'public, max-age=60, must-revalidate');
    else if (/sdk-[a-f0-9]{8,}\.js$/.test(filePath)) res.setHeader('Cache-Control', 'public, max-age=31536000, immutable');
    // Images, fonts, and WASM rarely change between deploys — cache 1 day.
    else if (filePath.match(/\.(svg|png|jpg|jpeg|webp|gif|ico|woff2?|wasm)$/)) res.setHeader('Cache-Control', 'public, max-age=86400, must-revalidate');
    // CSS/JS change with deploys — 1 hour with revalidate keeps deploys fresh
    // while removing the 5-minute thrash that was hammering edge caches.
    else if (filePath.match(/\.(css|js|map)$/)) res.setHeader('Cache-Control', 'public, max-age=3600, must-revalidate');
  },
}));

// RS-1 schema bundle — canonical JSON Schemas + spec markdown live in /docs
// so the homepage anchors (/docs#rs-1, #manifest, #receipts) and direct
// schema fetches both work. We mount the directory at /docs-static so the
// /docs SPA route below can still own the HTML page; specific filenames
// are then aliased back into /docs/* via explicit routes.
const DOCS_DIR = path.join(__dirname, 'docs');
for (const name of ['manifest-v0.1.json', 'receipt-v0.1.json', 'rs-1.md']) {
  app.get('/docs/' + name, (_req, res) => {
    const file = path.join(DOCS_DIR, name);
    if (!fs.existsSync(file)) return res.status(404).json({ error: 'spec asset not found' });
    if (name.endsWith('.json')) res.type('application/schema+json');
    else if (name.endsWith('.md')) res.type('text/markdown');
    res.set('Cache-Control', 'public, max-age=300');
    res.sendFile(file);
  });
}

app.use('/', buildRouter());

// SPA fallback for HTML routes — every public page maps to a static file under /public.
// /compile, /run, /recall, /cloud, /manual, /mobile are the v5 (`kolm`) surfaces.
// Legacy v4 pages (/optimize, /audit, /why, /how-it-works, /economics, /spec,
// /receipts, /verified, /specialists) stay reachable until Sprint 1's kill-list
// pass — the static files still live in public/.
// Page aliases — same file serves multiple URLs.
//  /signin → public/signup.html (one page, two tabs)
//  /atlas  → public/registry.html (registry → atlas rename, /registry kept for back-compat)
const ROUTE_ALIASES = {
  '/signin': 'signup',
  '/atlas': 'registry',
  '/spec/grammar': 'spec-grammar',
  '/developers': 'build-your-own',
  '/solutions': 'use-cases/index',
  '/teams/accept': 'teams-accept',
  '/signin': 'signup',
  '/login': 'signup',
};
for (const route of ['/', '/dashboard', '/playground', '/docs', '/registry', '/atlas', '/signup', '/signin', '/login', '/why', '/pricing', '/status', '/account', '/how-it-works', '/device', '/compile', '/run', '/recall', '/cloud', '/k-score', '/benchmarks', '/compare', '/research', '/serve', '/evolve', '/anatomy', '/security', '/privacy', '/terms', '/healthcare', '/finance', '/legal', '/edge', '/cookbook', '/defense', '/manifesto', '/faq', '/quickstart', '/trust', '/integrations', '/press', '/vs-ollama', '/vs-rag', '/vs-fine-tune', '/vs-predibase', '/vs-openpipe', '/vs-langsmith', '/vs-mem0', '/vs-hindsight', '/vs-openai-fine-tune', '/vs-together', '/why-now', '/threat-model', '/roi', '/api', '/whitepaper', '/build-your-own', '/developers', '/solutions', '/audit-log', '/baa', '/captures', '/capture', '/enterprise', '/glossary', '/leaderboard', '/spec', '/spec/grammar', '/models', '/compute', '/troubleshooting', '/teams', '/teams/accept', '/tunnels', '/byoc', '/airgap', '/showcase', '/sdks', '/compliance-packs']) {
  app.get(route, (_req, res) => {
    const name = route === '/' ? 'index' : (ROUTE_ALIASES[route] || route.slice(1));
    const file = path.join(__dirname, 'public', name + '.html');
    if (fs.existsSync(file)) return res.sendFile(file);
    res.status(404).sendFile(path.join(__dirname, 'public', '404.html'));
  });
}

// Extensionless use-case URLs: /use-cases/<slug> → public/use-cases/<slug>.html
app.get('/use-cases/:slug', (req, res, next) => {
  const slug = req.params.slug;
  if (!/^[a-z0-9-]+$/.test(slug)) return next();
  const file = path.join(__dirname, 'public', 'use-cases', slug + '.html');
  if (fs.existsSync(file)) return res.sendFile(file);
  next();
});

// Extensionless integration URLs: /integrations/<slug> → public/integrations/<slug>.html
app.get('/integrations/:slug', (req, res, next) => {
  const slug = req.params.slug;
  if (!/^[a-z0-9-]+$/.test(slug)) return next();
  const file = path.join(__dirname, 'public', 'integrations', slug + '.html');
  if (fs.existsSync(file)) return res.sendFile(file);
  next();
});

// Extensionless article URLs: /articles/<slug> → public/articles/<slug>.html
app.get('/articles/:slug', (req, res, next) => {
  const slug = req.params.slug;
  if (!/^[a-z0-9-]+$/.test(slug)) return next();
  const file = path.join(__dirname, 'public', 'articles', slug + '.html');
  if (fs.existsSync(file)) return res.sendFile(file);
  next();
});

// Extensionless research-article URLs: /research/<slug> → public/research/<slug>.html
app.get('/research/:slug', (req, res, next) => {
  const slug = req.params.slug;
  if (!/^[a-z0-9-]+$/.test(slug)) return next();
  const file = path.join(__dirname, 'public', 'research', slug + '.html');
  if (fs.existsSync(file)) return res.sendFile(file);
  next();
});

// /cookbook/<slug>: serves recipes from public/cookbook/<slug>.html if present.
// Vertical aliases (healthcare/finance/legal/edge) keep their canonical /<vertical> URLs;
// /cookbook/<vertical> serves the same file so either path works.
const COOKBOOK_VERTICALS = new Set(['healthcare', 'finance', 'legal', 'edge']);
app.get('/cookbook/:slug', (req, res, next) => {
  const slug = req.params.slug;
  if (!/^[a-z0-9-]+$/.test(slug)) return next();
  if (COOKBOOK_VERTICALS.has(slug)) {
    const file = path.join(__dirname, 'public', slug + '.html');
    if (fs.existsSync(file)) return res.sendFile(file);
  }
  const file = path.join(__dirname, 'public', 'cookbook', slug + '.html');
  if (fs.existsSync(file)) return res.sendFile(file);
  next();
});

// 404 fallback for unknown HTML routes — branded page from /public/404.html if it exists.
const _404Path = path.join(__dirname, 'public', '404.html');
app.use((req, res, next) => {
  if (req.method === 'GET' && req.accepts('html') && !req.path.startsWith('/v1') && !req.path.startsWith('/health') && req.path !== '/ready' && req.path !== '/404') {
    if (fs.existsSync(_404Path)) return res.status(404).sendFile(_404Path);
    return res.status(404).type('html').send(`<!DOCTYPE html><html><head><title>404 · kolm</title><link rel="stylesheet" href="/styles.css"></head><body style="padding:48px;text-align:center;font-family:system-ui;color:#e8ecf3;background:#0a0b0e;min-height:100vh;"><h1 style="font-size:48px;margin:0;letter-spacing:-0.02em;">404</h1><p style="color:#8b94a8;margin-top:8px">That page doesn't exist.</p><p style="margin-top:24px;"><a href="/" style="color:#7dd3fc;">&larr; Home</a> &middot; <a href="/registry" style="color:#7dd3fc;">Registry</a> &middot; <a href="/docs" style="color:#7dd3fc;">Docs</a></p></body></html>`);
  }
  next();
});

// Generic 500 — catches any unhandled error in routes. Standardized shape:
// `{ error, detail? }` where `detail` is omitted in production to avoid
// leaking stack-tail strings or internal paths.
app.use((err, req, res, _next) => {
  console.error('[500]', err);
  if (req.accepts('html')) {
    const _500Path = path.join(__dirname, 'public', '500.html');
    if (fs.existsSync(_500Path)) return res.status(500).sendFile(_500Path);
  }
  const body = { error: 'internal server error' };
  if (process.env.NODE_ENV !== 'production' || process.env.KOLM_DEBUG) {
    body.detail = String(err.message || err);
  }
  res.status(500).json(body);
});

const PORT = parseInt(process.env.PORT || '8787');

async function bootSeedDemoConcepts(tenant) {
  const dir = path.resolve('examples');
  if (!fs.existsSync(dir)) return { added: 0, skipped: 0 };
  const files = fs.readdirSync(dir).filter(f => f.endsWith('.json'));
  const existing = new Set(all('concepts').filter(c => c.tenant === tenant).map(c => c.name));
  let added = 0, skipped = 0;
  for (const file of files) {
    try {
      const ex = JSON.parse(fs.readFileSync(path.join(dir, file), 'utf8'));
      if (existing.has(ex.name)) { skipped++; continue; }
      const r = await synthesize({ positives: ex.positives, negatives: ex.negatives || [], output_spec: ex.output_spec, priors: ex.priors || {} });
      if (!r.accepted) { skipped++; continue; }
      const concept = createConcept({
        name: ex.name, description: ex.description || ex.name, tenant,
        schema: ex.output_spec || null, tags: ex.tags || [], visibility: ex.visibility || 'public',
      });
      publishVersion({
        concept_id: concept.id, source: r.source,
        evaluation: { quality_score: r.quality_score, pass_rate_positive: r.pass_rate_positive, reject_rate_negative: r.reject_rate_negative, latency_p50_us: r.latency_p50_us, size_bytes: r.size_bytes, source_hash: r.source_hash, strategy: r.strategy, trace: r.test_trace },
        lineage: { synthesized_from_n: ex.positives.length + (ex.negatives?.length || 0), attempts_n: r.attempts_n },
      });
      added++;
    } catch { skipped++; }
  }
  return { added, skipped };
}

if (process.argv[1] && process.argv[1].endsWith('server.js')) {
  // Auto-provision the demo tenant.
  const demo = provisionTenant(process.env.DEFAULT_TENANT || 'demo');

  // Idempotent seed: synthesizes any missing example/*.json concepts.
  const { added, skipped } = await bootSeedDemoConcepts(demo.name);
  if (added > 0 || skipped > 0) console.log(`  seed: +${added} added, ${skipped} skipped`);

  // Stripe configuration sanity check. Counts how many of the 5 payment links
  // are wired so a misconfigured deploy logs a single line at boot rather than
  // surfacing as a 503 the first time a customer clicks Upgrade. Webhook secret
  // is required for paid plans to flip; surface the gap explicitly.
  const stripeLinks = [
    'STRIPE_PAYMENT_LINK_STARTER', 'STRIPE_PAYMENT_LINK_PRO',
    'STRIPE_PAYMENT_LINK_TEAMS', 'STRIPE_PAYMENT_LINK_BUSINESS',
    'STRIPE_PAYMENT_LINK_ENT',
  ];
  const stripePresent = stripeLinks.filter(v => !!process.env[v]).length;
  const webhookOk = !!process.env.STRIPE_WEBHOOK_SECRET;
  const stripeStatus = stripePresent === 5 && webhookOk ? 'wired' : `degraded (${stripePresent}/5 links, webhook ${webhookOk ? 'ok' : 'missing'})`;

  app.listen(PORT, () => {
    console.log('\nkolm server');
    console.log(`  home:       http://localhost:${PORT}`);
    console.log(`  dashboard:  http://localhost:${PORT}/dashboard`);
    console.log(`  playground: http://localhost:${PORT}/playground`);
    console.log(`  docs:       http://localhost:${PORT}/docs`);
    console.log(`  demo key:   ${!!demo.api_key ? 'configured' : 'missing'}`);
    console.log(`  admin key:  ${!!process.env.ADMIN_KEY ? 'configured' : 'not set'}`);
    console.log(`  synthesis:  ${process.env.ANTHROPIC_API_KEY ? 'Claude (' + (process.env.ANTHROPIC_MODEL || 'claude-opus-4-7') + ') + Pattern' : 'Pattern (no API key set)'}`);
    console.log(`  stripe:     ${stripeStatus}`);
    console.log('');
  });
}

export { app };
export default app;